pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "adb-agent"
version = "0.1.0"
description = "Phone automation agent using Google ADK + Android ADB with pure vision approach"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=2.0.0a0",
    "httpx[socks]>=0.28.1",
    "Pillow>=10.0.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.30.0",
]

[project.scripts]
adb-agent = "main:main"
browser-agent = "browser_main:main"
adk-web-agents = "adk_web_agents.cli:main"
scraper-agent = "scraper_main:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.setuptools]
py-modules = ["main", "browser_main"]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["adb_agent*", "browser_agent*", "agent_shared*", "adk_web_agents*", "scraper_agent*"]
exclude = ["tests*"]
